refactor: updated the component where we fetch the group data
